+++ ChannelIO/Source/Controllers/UserChatsViewController.swift
@@ -10,6 +10,7 @@ import UIKit
 import CHDwifft
 import ReSwift
 import RxSwift
+import RxSwiftExt
 import SnapKit
 import SVProgressHUD
 import MGSwipeTableCell
@@ -245,47 +246,41 @@ class UserChatsViewController: BaseViewController {
     self.isShowingChat = true
     
     let controller = self.prepareUserChat(userChatId: userChatId, text: text)
-    let channel = mainStore.state.channel
     
     //NOTE: Make sure to call onCompleted on observable method to avoid leak
-    let pluginId = mainStore.state.plugin.id
-    let pluginSignal = CHPlugin.get(with: pluginId)
+    let pluginSignal = CHPlugin.get(with: mainStore.state.plugin.id)
     let followersSignal = CHManager.getRecentFollowers()
-    let supportBot = channel.canUseSupportBot ?
-      CHSupportBot.getBots(with: pluginId, fetch: userChatId == nil) :
-      .just([])
     
-    var pluginBot: CHBot? = nil
+    var bot: CHBot? = nil
     
-    Observable.zip(pluginSignal, followersSignal, supportBot)
-      .observeOn(MainScheduler.instance)
-      .flatMap({ (info, managers, supportBots) -> Observable<CHSupportBotEntryInfo> in
-        mainStore.dispatch(UpdateFollowingManagers(payload: managers))
-        mainStore.dispatch(GetPlugin(plugin: info.0, bot: info.1))
-        
-        pluginBot = info.1
-        mainStore.dispatch(GetSupportBots(payload: supportBots))
-        //target evaluate (supportBots) -> bot? 
-        //evaluation happen here later
-        
-        if let supportBot = supportBots.first {
-          return supportBot.getEntry()
-        } else {
-          return .just(CHSupportBotEntryInfo(step: nil, actions: []))
-        }
+    Observable.zip(pluginSignal, followersSignal)
+      .retry(.delayed(maxCount: 3, time: 3.0), shouldRetry: { error in
+        SVProgressHUD.show()
+        dlog("Error while opening chat. Attempting to open again")
+        return true
+      })
+      .flatMap({ (info, managers) -> Observable<CHSupportBotEntryInfo> in
+        bot = info.1
+        mainStore.dispatchOnMain(UpdateFollowingManagers(payload: managers))
+        mainStore.dispatchOnMain(GetPlugin(plugin: info.0, bot: info.1))
+        return CHSupportBot.get(with: mainStore.state.plugin.id, fetch: userChatId == nil)
+      })
+      .catchError({ (error) -> Observable<CHSupportBotEntryInfo> in
+        return .just(CHSupportBotEntryInfo())
       })
       .observeOn(MainScheduler.instance)
-      .subscribe(onNext: { [weak self] (entryInfo) in
-        if entryInfo.step != nil {
-          mainStore.dispatch(GetSupportBotEntry(bot: pluginBot, entry: entryInfo))
+      .subscribe(onNext: { [weak self] (entry) in
+        SVProgressHUD.dismiss()
+        if entry.step != nil && entry.supportBot != nil {
+          mainStore.dispatch(GetSupportBotEntry(bot: bot, entry: entry))
         }
-        
         self?.navigationController?.pushViewController(controller, animated: animated)
         self?.showNewChat = false
         self?.isShowingChat = false
         self?.shouldHideTable = false
         dlog("got following managers")
       }, onError: { [weak self] (error) in
+        SVProgressHUD.dismiss()
         dlog("error getting following managers: \(error.localizedDescription)")
         self?.showNewChat = false
         self?.isShowingChat = false
@@ -462,11 +457,15 @@ extension UserChatsViewController {
     }
     
     UserChatPromise.getChats(since: isInit ? nil : self.nextSeq, limit: 30, showCompleted: self.showCompleted)
+      .retry(.delayed(maxCount: 3, time: 3.0), shouldRetry: { error in
+        dlog("Error while fetching chat data. Attempting to fetch again")
+        return true
+      })
       .observeOn(MainScheduler.instance)
       .subscribe(onNext: { [weak self] (data) in
-        mainStore.dispatch(GetUserChats(payload: data))
-        
         self?.didLoad = true
+        
+        mainStore.dispatch(GetUserChats(payload: data))
         self?.showChatIfNeeded(isReload: isReload)
       }, onError: { [weak self] error in
         dlog("Get UserChats error: \(error)")
@@ -499,7 +498,7 @@ extension UserChatsViewController {
   }
   
   func showChatIfNeeded(_ userChats: [CHUserChat]? = nil, isReload: Bool = false) {
-    let allChats = userChatsSelector(state: mainStore.state)
+    let allChats = userChatsSelector(state: mainStore.state, showCompleted: self.showCompleted)
     
     if self.showNewChat  {
       self.showUserChat(animated: false)